Major refactoring of NLog2 code base:
This is intermediate checkin. Code is not guaranteed to be of any quality at this moment and unit tests are not enabled yet.
Refactored build process to produce .NET 2.0/3.5, .NET CF 2.0/3.5 and Silverlight 2 version of the library.
Added StyleCop to the build process - fixed 13000 warnings which unified the codebase.
Beginnings of a WIX installer, Documentation generation via Sandcastle.