Add priority to MailTarget
Currently all messages are sent with normal priority, it would be nice to
be able to sent fx. Fatal log messages with high priority.
To solve this, Priority has been added to the MailTarget using a Layout,
this allows users to define the priority based on conditions in the log
message fx. log level.
fixes #198